requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
abseil-cpp=20210324.2=h9c3ff4c_0
absl-py=1.0.0=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py39hb9d737c_2
async-timeout=4.0.2=pyhd8ed1ab_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.4.26=h06a4308_0
cached-property=1.5.2=hd8ed1ab_1
charset-normalizer=2.0.12=pyhd8ed1ab_0
dm-tree=0.1.7=pypi_0
flit-core=3.7.1=pyhd8ed1ab_0
google-auth=1.35.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpc-cpp=1.42.0=ha1441d3_1
importlib-metadata=4.11.3=py39hf3d152e_0
jupyterlab-widgets=1.1.0=pypi_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
ld_impl_linux-64=2.35.1=h7274673_9
libgcc-ng=11.2.0=h1d223b6_14
libgfortran-ng=11.2.0=h69a702a_14
libstdcxx-ng=11.2.0=he4da1e4_14
lz4-c=1.9.3=h295c915_1
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mir-eval=0.7=pypi_0
more-itertools=8.13.0=pypi_0
nbconvert-core=6.5.0=pyhd8ed1ab_0
nbconvert-pandoc=6.5.0=pyhd8ed1ab_0
nest-asyncio=1.5.4=pyhd8ed1ab_0
notebook-shim=0.1.0=pyhd8ed1ab_0
onnx-tf=1.10.0=pypi_0
prompt-toolkit=3.0.27=pyha770c72_0
pyasn1-modules=0.2.7=py_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python-graphviz=0.20=pypi_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
scikit-learn=1.0.2=pypi_0
tensorboard-data-server=0.6.0=py39h95dcef6_1
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow-addons=0.17.0=pypi_0
tensorflow-base=2.7.0=cuda112py39hd98b2dd_0
tensorflow-estimator=2.7.0=cuda112py39h23446aa_0
tensorflow-gpu=2.7.0=cuda112py39h0bbbad9_0
tensorflow-probability=0.13.0=pypi_0
typing-extensions=4.1.1=hd8ed1ab_0
websocket-client=1.3.2=pyhd8ed1ab_0
